line_chatbot_genai/requirements.txt
requests
line-bot-sdk==3.13.0
